requirements.txt
fastapi>=0.45.0
uvicorn>=0.17.0
mysql-connector-python>=8.0.28